If the log doesn't fit in the float, something is terribly wrong
This one bombed out:

params = LWE.Parameters(n=700, q=next_prime(2^13), Xs=ND.SparseTernary(700, 100), Xe=ND.UniformMod(11))
LWE.dual_hybrid(params)

+++ estimator/prob.py
@@ -92,10 +92,10 @@ def amplify(target_success_probability, success_probability, majority=False):
 
     prec = max(
         53,
-        2 * ceil(abs(log(success_probability, 2))),
-        2 * ceil(abs(log(1 - success_probability, 2))),
-        2 * ceil(abs(log(target_success_probability, 2))),
-        2 * ceil(abs(log(1 - target_success_probability, 2))),
+        2 * ceil(abs(float(log(success_probability, 2)))),
+        2 * ceil(abs(float(log(1 - success_probability, 2)))),
+        2 * ceil(abs(float(log(target_success_probability, 2)))),
+        2 * ceil(abs(float(log(1 - target_success_probability, 2)))),
     )
     prec = min(prec, 2048)
     RR = RealField(prec)
